Remove "phys_area" from output in HOMME template test
[BFB]
This variable is not needed, and was causing failures at the roundoff level: it does not
appear to be BFB accross multiple runs. The way it is calculated must
not be MPI safe.  But this is for a feature of HOMME not used by ACME so removing
this for now.

*********1*********2*********3*********4*********5*********6*********7**
Longer commit message body describing the commit.
Can contain lists as follows:
	* Item 1
	* Item 2
	* Item 3

A good commit message should be written like an email, a subject
followed by a blank line, followed by a more descriptive body.

Can also contain a tag at the bottom describing if the commit is
Non-BFB or Climate changing:
[Non-BFB]
[CC]