Changing velocity interpolation in output.
*********1*********2*********3*********4*********5*********6*********7**
Introducing different interpolation of velocity fields in output. The previous interpolation used contravariant bases which are not continuous across elements' boundaries. The new code transforms (u,v) in lonlat basis to Cartesian basis, interpolates each of three components as a scalar, and transforms the result back to the lonlat basis.

The modified file is interpolate_mod. The routines affected are interpolate_vector2d and interpolate_vector3d. The first routine is not verified since now shallow water code uses velocities with level dimension. The previous version of the interpolation is not erased, it is disabled with if(.FALSE.).

These changes are BFB and will only affect plots and nc outputs, though it is unlikely that users would see any difference.

[BFB]