Mod in cmake/Trilinos.cmake
Adding a verbose comment to tell the user that we didn't find Trilinos,
and informing him/her about the TRILINOS_INSTALL_DIR variable.

Also, the default path for TRILINOS_INSTALL_DIR is now inside the build
tree, to avoid polluting the user's home directory. Furthermore, source,
build and install folder will all be in the build tree, under the folder
external/trilinos, to avoid polluting the source tree of acme with the
source folder of trilinos.