Added a new file: PhysicalConstants.*cpp
The new structure contains all the constants used in homme, and fulfills
the same role that the physical_constants module does in F90.

The free function get_physical_constants_c() provides a static instantiation
of this type, allowing to access the physical constants everywhere in the
code. Before doing so, though, one has to initialize the constants stored
in this static instantiation, using the free function init_physical_constants_c().
This is done in the sweqx module.