Added a new file: ControlParameters.*cpp
The new structure contains some of the simulation parameters used in homme,
and fulfills the same role that the control module does in F90.

The free function get_control_parameters_c() provides a static instantiation
of this type, allowing to access the simulation parameters everywhere in the
code. Before doing so, though, one has to initialize the parameters stored
in this static instantiation, using the free function init_control_parameters_c().