Added a new file: PointersPool.*pp
This file defines a structure which contains a variety of pointers, which
are needed in different parts of the code to properly initialize Kokkos
views. The free function 'get_pointers_pool_c()' returns a static instance
of this type, much as a global variable. This instance can then be used to
access the needed pointers anywhere in the code. Before doing so, though,
the user has to properly initialize the static instance using the free
function 'init_pointers_pool_c()'.

Note: the rational behind this struct is to avoid having really long interfaces
for the kokkos kernels. In fact, in F90 one can simply use a module where
a pointer is defined in order to access it. This is not possible in C/C++,
so in order to avoid passing all the pointers around, we can use a static
instance of a struct which stores all of them.