New file: kokkos_kernels/SphericalOperators.*pp
This file contains the kokkos-ized version of all the routines that are in
the derivative module in F90. All the kernels have a duplicated interface:
one version wher they accept raw pointers, and one where they accept Kokkos
views (the former simply creates views and invokes the latter). I'm not sure
if they are all needed.

Note: in order to exploit Kokkos subviews, I had to use LayoutStride in
almost every single view. This could reduce performance, compared to LayoutLeft
and LayoutRight. We may have to investigate this further.