Added unit tests for laplace_sphere-related kernels
I had to also create a fortran module, where the derivative and element
structures could be initialized in F90. This is because the F90 routines
that compute laplace stuff extensively use element/derivative structures,
and it would take too much effort to convert them to use flat arrays just
for the sake of testing.