Configure fix in Trilinos.cmake
The variables Kokkos_LIBRARIES and Kokkos_TPL_LIBRARIES should not have the
PARTENT_SCOPE option, since the Trilinos.cmake is invoked through an INCLUDE
directive. INCLUDE and MACRO do not open a new scope. Since Trilinos is
included in the main CMakeLists file, this would generate a cmake warning
saying that there is no parent scope. Hence, the variable would not be set.
In case this is the first build (where we also build trilinos), the Trilinos
cmake config file would not be imported at config time, and we would end up
with Kokkos_LIBRARIES and Kokkos_TPL_LIBRARIES being empty strings, giving
us linker errors. Everything would work fine on our second build (when the
Trilinos cmake config file would be finally imported, defining those two
variables).

With this fix, the hard coded values for the libraries are available at
configure time if Trilinos is not yet installed.