Mod to Trilinos.cmake
Not building Trilinos in the build directory. It is too cumbersome if we
have many different build trees. However, we do keep the debug and release
installation of trilinos separated in ~/prefix/trilinos, in case we need
to use a debug version of trilinos (for whatever reason).

However, unless the user specifically requests it, we build trilinos in
RELEASE mode.