Checking in scripts for nightly testing on ride with a CUDA KokkosNode. Testing will be activated tonight and results will be posted to this CDash site on the SRN:
http://cdash.sandia.gov/CDash-2-3-0/index.php?project=HOMMEXX

In the nightly testing, the ctest results are compared against baselines created today in my
home directory.  All tests pass except test13, which I believe Michael
is working to fix in an alternate branch.

Note that as a part of the ride testing, only HOMMEXX is built, against
an existing Trilinos in /home/mdeakin/prefix.  The cmake rules
should be changed in HOMMEXX to enable building a working Trilinos
inside the code.  Need to talk to Michael about this.