Changing Views types names plus bug fix in advance_rk.cpp
First, we define the aliases HostSpace, ExecSpace and ScratchSpace, and
the two memory management types MemoryManaged and MemoryUnmanaged.

Homme defines a general alias to Kokkos::View, which accept four template
parameters: DataType, MemorySpace, MemoryManagement and LayoutType. The
latter defaults to LayoutLeft. In addition, other types are defined:

1) HommeView1D,...,HommeView6D: these types are partial specializations of
   the main HommeView type, with DataType = real*,....,real****** respectively.
2) HommeHostView1D,...,HommeHostView6D: these types are partial specializations
   of the ones at 1), with MemorySpace = HostSpace. The memory management
   remains as a template parameter, since we can have views created from
   pointers (unmanaged) or not (managed).
3) HommeExecView1D,...,HommeExecView6D: these types are full specializations
   of the ones at 1), with MemorySpace = ExecSpace and
   MemoryManagement = MemoryManaged.
4) HommeScratchView1D,...,HommeScratchView6D: these types are full specializations
   of the ones at 1), with MemorySpace = ScratchSpace and
   MemoryManagement = MemoryUnmanaged.

If by any chance you need an unmanaged view on the ExecSpace or a managed
view on the Scratch space, you need to use HommeView1D,...,HommeView6D,
with the proper template arguments.

As for the bug, there were some views in advance_rk which were 2D but were
accessed with 3 indices. For some reason Kokkos does not issue a compiler
error nor gives a run-time error if the user passes the wrong number of
arguments either at construction time or to the operator() (not even in
debug mode!).