Starting to setup the preqx_flat testing.
I have started to split the compute_and_apply_rhs routine in prim_advcance_mod into smaller
kernels, to allow incremental conversion to kokkos (one little kernel at
a time). However, for now it does not work. I left an #ifdef 0 in the code
which excludes the new reorganization of the code from being compiled.

Some changes required to also change a few things back in sweqx.

This commit is a sort of backup, to allow to share the current work. Many
things have to be fixed still. For instance, if USE_KOKKOS_KERNELS is defined,
the code won't compile, since the c counterpart of the caar routines are yet
to be written.