I continued working on breaking down compute_and_apply_rhs into smaller routines.
There is still something off. After the first stage of caar, I obtain the same
velocity and temperature but not dp3d (this is comparing the original version,
with a single big loop, with the new version, with the loop broken down into
smaller pieces).

I left a bunch of print statements in the code, to allow visualization, and
I abort after the first stage of caar, since that helps with visualization
(no point going forward when we know that there's already a bug there).

The bug is really subtly hidden: the two versions compute the same local
dp3d, but the exchange part and the subsequent unpack screw up something.
I suspect that a pointer is not set correctly somewhere, since I refuse to
believe the bug is inside the exchange subroutines...

Note: to run the original version, edit the config.h file in the build folder,
and add

#define CAAR_ORIG