Ah-ah! Found the bug and fixed it. Now the two versions give the same results.
The bug was indeed hidden: I was calling c_f_pointer passing the wrong
dimensions for the fortran pointer. This would be caught if the f-pointer
was accessed with indices. However, the code looked like this

! Note: a_c_ptr points to a [2, 3] array)
call c_f_pointer(a_c_ptr, a_f_ptr, [2, 2])

a_f_ptr = 0

So, although the dimensions of a_f_ptr were wrong, fortran could not now it.
This caused some entries in a_c_ptr to not be initialized to 0.

Note: this was on eta_dot_dpdn, which has nlev+1 levels, not nlev. WATCH OUT FOR THIS KIND OF BUGS!