Created subfolder share/cxx for cxx stuff which is fairly common among homme
This includes dimensions, types and utility. Also, moved kokkos_init in here,
and added an integer parameter, the number of threads. I don't know how to
easily read command linke inputs in f90 and forward them to kokkos (to allow
--kokkos-xyz=blah). However, we can manually pass the number of threads to
the init function. This is actually better (for now), since the number of
threads for vertical parallelization is decided by homme, looking at the
namelist file.