Started porting kokkos kernels.
I ported only one for now, but it took
some effort, due to the fact that, in order to both use f90 ptrs and keep
compile time dimensions, we need to manually flip the views before and after
the calculations. This adds an overhead, which will get smaller and smaller
as we merge kernels together, and eventually disappear when we have only
cxx code.