Added a macro to generate config file only if it would be different from before.
Since we keep changing CMakeLists.txt files, we force cmake to be re-run.
This, in turn, would change the config files. Since these are included in
a lot of files, we end up with lots of files rebuilt even if nothing changed.

With this macro, we create a temporary config file and we compare it with the
existing one (if any). If they compare equal, simply delete the temporary
config file and we're done. If they compare different or there was no
config file to start with, then we create the configure files as usual.