Spherical operators file.
So far it only implements gradient_sphere (two versions), divergence_sphere,
and vorticity_sphere. More functions may follow. Also, the functions may
be turned into methods of a struct, if deemed more efficient.

Also notice that, as of now, the functions operate on NPxNP views. That is,
to compute, say, the gradient of a NLEVxNPxNP field, you need to loop on the
levels, and call gradient_sphere on a subview on each level. Also, the
functions assume that they are called from inside a parallel_for, with
team policy.