Finished refactor of second kernel. Also, removed the cxx utility that flips views, and moved it to f90 code.
The flip from f90 to cxx ordering is now done inside the f90 code. This is
not super efficient, since in cxx we could do it on the execution space in
parallel. However, that would cause us to include the flipping time in the
timers, and we would pollute the performance timings. By having f90 do the
conversion, we can start/stop the clocks (t_startf/t_stopf) after/before the
conversion, so that we time just the kernels.

I also created a help function for deep_copy. The reason is the following.
Right now, Kokkos::deep_copy has a compile time check to ensure that the
data type of the destination view is NOT const. This can be a problem if
you create view_out using create_mirror_view, passing the output memory
space. In particular, if you have a View view_in with const data type on
memory space MS, and you create a mirror view on the same memory space
(perhaps you're using OpenMP), the output view will be a shallow copy of
view_in, with also const data type. If you then try call deep_copy
on these two views, the compiler will bite you.