Removed unused typedefs for some views and added another double* typedef.
The new typedef F90Ptr is the same as CRCPtr, that is 'const double* const'.
The reason for having another name for the same type is that this name
stresses the fact that the given pointer is pointing to a fortran array.
In other words, if I see 'F90Ptr p', I know that that pointer was created,
somewhere in some F90 code, with something like 'p = c_loc(array)'.