A header containing a wrapper for some array/routines coming from derivative_mod
The arrays are stored in views inside a Derivative structure. So far only
three arrays made it in the class, but more can be added. The user can
access a static instance of the class via the get_derivative. The static
instance must be initialized calling init_derivative_c (which is callable
from fortran), passing the pointers to the corresponding fortran arrays.

The only routine now in this header is subcell_div_fluxes.

Note: we had to do a little mod to derivative_mod, namely making two arrays public and target.