Removed restriction on memory space for compute_host_view_norm
Coherently, I changed the name into compute_view_norm. We don't care where
the view resides, as long as we can access the data. I can't think of a way
to detect whether we can access the view data from where we are, so do not
call this function on the host if the view is on the device (with a different
memory space) and viceversa.