Bug fix relative to qn0 in the kokkos routines.
Although we change all fortran indices to the C 0-based standard, when the
qn0 index is set to -1 in fortran, we should leave it like that, since it
is not an index (per se), but it signal that something else should be done.
The alternative would be to change the if(qn0==-1) to if (qn0==-2) in C,
but -1 is more common as "invalid" index.