Printing the sha of the last git commit at configure time
I added this because I saw some errors on the CDash output that should have
been fixed by a commit pushed the day before the CDash ran. To make sure
we know what revision we are testing, we can take a look to the sha of the
last git commit, printed at configure time (right after all compilers
detection routines).