Create preqx_flat testing
There are three tests ran for a given configuration:

 1a) the f90 code is ran and output checked against baseline
 1b) the kokkos code is ran and output checked against baseline
 2)  the f90 and kokkos outputs are checked against each other

Notice that, in general, the f90 and kokkos baselines should match, so
the third test is almost superfluous. But in case the first two check do
not pass, the third one can at least tell us if f90 and kokkos versions
match.

I had to introduce some cmake machinery for step 2). To add another test,
first proceed as usual, adding the cmake files for f90 and c versions
in test/reg_test/run_tests, and add the tests to the HOMME_TESTS list in
test-list.cmake file. To add the step 2 above, add the test root name to
the HOMMEXX_TESTS list in test-list.cmake.
WARNING: the implementation STRONGLY assumes that the two folders for the
f90 and kokkos tests have names like the following:
 - my_test  : contains f90 test
 - my_test_c: contains kokkos test
Any change to this structure will break the testing. A workaround for this
would be to have two separate lists, like HOMME_TESTS_F90 and HOMME_TESTS_CXX,
containing the names of the tests to compare against each other. I may
actually implement this change in the next few days. Notice that, even in
this case, you should at least order corresponding tests in the same order
in the two lists...