Removed unused routines from SphereOperators
I removed the routines that do not require the user to pass in the views
to store temporaries. These routines created a temporary on the fly, and
passed it to the routines that required the temporary.

Since no other routine was using these (they where used by the non monolithic
caar routines), I decided to keep the code base smaller and removed them.