Added qsize to the control structure
We need this parameter for the tracers. In fact, despite the existence of
the macro QSIZE_D, one can apparently pass the number of tracers via the
namelist (it must not exceed QSIZE_D). QSIZE_D serves then as an upper
bound, but one can run homme with any number of tracers in [0,QSIZE_D].

This is an issue when copying the results of the euler step back to F90,
since the Qtens variable is dimensioned on qsize (passed through namelist)
and not on QSIZE_D (specified at compile time), and therefore may have less
than QSIZE_D tracers.