Changed the default value of AVX_VERSION to 'None'.
With '0', then if(NOT AVX_VERSION) would evaluate to true, while instead
I would like to check whether the user set AVX_VERSION or not. Now, if we
see it's left to 'NONE', we assume it was not set by the user. The user
should indeed set it to '0' if no AVX is requested.