Bug fix for team size handling in Control.
The previous implementation was setting default to 99999 then taking the min
with the value provided by DefaultThreadsDistribution. This would always lead
to a team_size=1 in case KOKKOS_COLUMN_THREAD_ONLY was not defined, even if
the user explicitly requested a larger team size (via env variable). This
commit fixes this.