Starting to refactor the views layout, to make it tiled.
The idea is to move the data type from, say, Scalar*[NP][NP][NUM_LEV] to
Scalar*[NUM_LEV][NP][NP]. Notice that this is not the same as the flipped
version of the fortran layout (np,np,nlev,nelemd), since the data type
Scalar still has VECTOR_SIZE doubles in it. This is why we refer to this
layout as tiled.