KNL optimizations.
This branch isn't intended to be merged. Rather, it was a sprint to
discover and document a number of changes we can make to improve
performance on KNL. These can be brought in incrementally:

1. Don't use subview unless the cost of construction is sufficiently
amortized.

2. Don't construct or copy-construct Views at each dispatch.

3. Reorder [pack-level, SIMD-level, (i,j)] level integration loops to
[pack-level, (i,j), SIMD-level] for locality.

4. Use refs to vector packs where possible to minimize indexing.

5. ThreadVectorRange doesn't seem always to get auto
vectorization. I've used #pragma simd in two spots with measurable and
substantial speedup. Unfortunately, this annotation ignores
-fp-model. It's a topic for discussion.

6. Although not turned on, I've stripped parallel_for on TeamPolicy
for the 1 element/rank case to show that there's avoidable dispatch
overhead.

More optimizations are possible, but they tend to be the same as
above, so I have not made all of them.

For ne=2, 4 days, with settings
    --bind-to core --map-by ppr:24:node:pe=2 -x OMP_NUM_THREADS=2 \
    -x KMP_AFFINITY=balanced
these changes bring time from 0.69s to 0.45, compared with 0.27 for
ACME master. So there's a ways still to go.