Removed explicit 0-initialization of Scalar's
Notice that our Scalar's default constructor already initializes to 0,
so there is no need for an extra loop. Moreover, for AVX vectors, the
default constructor uses _mm256_setzero_pd, while the constructor that
takes a double as input calls _mm256_set1_pd. Now, I don't know for sure
the performance difference of these two intrinsics, and it may as well
be that the compiler recognize that the input value is 0.0 and uses
_mm256_setzero_pd instead of _mm256_set1_pd. However, using the former
is for sure not slower than using the latter, and perhaps it buys us
some extra speedup.