Config: Fix quoted COMPILE_OPTIONS.
COMPILE_FLAGS nicely takes flag specifications with spaces in
them; COMPILE_OPTIONS does not. Instead COMPILE_OPTIONS produces a
compile line that *quotes* the option, e.g.: -O3 "-fp-model
strict", if one is not careful. Thus, use separate_arguments to
make a list, then pass the list to COMPILE_OPTIONS, which then
converts it to an unquoted space-delimited thing, which is what we
want.