Small bug fix for gcc.
The linker complains about an undefined reference to np1. The problem is
that, when a variable is odr-used (like passed by reference to a method,
such as operator() in Kokkos::View), then a definition (not just declaration)
must be present. An in-class constexpr declaration is not a definition, even
if the const value is specified. Somehow, intel allows this, but
gcc is strictly following the standard here.