Change in sphere operator, work in progress for PR #78.
Note: this commit maintains integrity of share_ut test and preqx_flat_sphere_op
test, but breaks preqx_flat. This is because now we need to change the
behavior of CaarFunctor, to reflect the fact that the sphere operators act
on the whole column of data, rather than just a level pack.