Thread device unification approach 1 for GPU.
At present we do not have a SIMD equivalent for GPU. This may change in the near
future, approx. 3 months.

In the meantime, we are considering a few different approaches. This branch
implements the most straightforward. It branches from thread_device_unification
so that we're not immediately committing to this approach.

First, in SphereOperators and phases 1 and 3, this approach replaces each serial
for ilev loop with ThreadVectorRange. Nothing further is needed here. It's not
clear to me that SIMD for GPU would necessarily do better, but it would be
clearer.

Second, in the scan properties phase, do more complicated stuff in the same
spirit. In this commit, I've simply run 1 thread (using
Kokkos::serial(Kokkos::PerThread(kv.team), ...)) from each vector pack and let
the rest do nothing. That means only 16 (for 16 GLL points) out of 128 threads
in a 4-warp thread block are being used. I think a little trickery should get us
back up to 128 threads, but that will be for future commits.

There are still a couple of broken parts of the unit tests, so this commit isn't
completely correct. But a quick look at timing suggests this commit brings
thread_device_unification to better than parity with master on GPU.