Adding a file ExecSpaceDefs, where to put all metaprogramming for the ExecSpace
This file defines Hommexx_XYZ, for XYZ={Serial,Threads,OpenMP,Cuda}. If the
execution space is enabled in Kokkos, Hommexx_XYZ=Kokkos::XYZ, otherwise
Hommexx_XYZ=void. Later, ExecSpace is set to alias one of Hommexx_XYZ, depending
on the user choice. This way we can simply test whether Hommexx_XYZ is = void
in case we need XYZ-specific code (usually only for XYZ=Cuda).

In this file we also do the DefaultThreadsDistribution metaprogramming magic,
to select the "best" configuration for the team size.