WIP EulerStepFunctor topic branch.
This branch will focus on the on-node work in euler_step. Probably it will be
merged into master when that is done. Then, once the C++ MPI branch is merged
into master, a new topic branch will focus on switching to the C++ MPI code. A
valuable conversation with Oksana and subsequent analysis shows the following
must be done:

    prim_advection_mod_base.F90
    active defines: OVERLAP NEWEULER_B4B _PRIM
    Prim_Advec_Tracers_remap_rk2
      euler_step
        ! compute Q min/max values for lim8
        ! compute biharmonic mixing term f
        if rhs_multiplier = 0
          neighbor_minmax
        elif rhs_multiplier = 2
          neighbor_minmax_start
          biharmonic_wk_scalar
          neighbor_minmax_finish
        ! 2D Advection step
        advance_qdp
        limiter_optim_iter_full
        bndry_exchangeV
    viscosity_mod_base.F90
      neighbor_minmax*: S(calar) bndry exchange + trivial min/max
      biharmonic_wk_scalar
        laplace_sphere_wk
        bndry_exchangeV
        laplace_sphere_wk
    derivative_mod_base.F90
      limiter_optim_iter_full: the cell-local QP solver

So far, this branch has implemented roughly the right parallel dispatch that
we'll need, but it has hacks and to-dos related to reworking
DefaultThreadsDistribution.