Added the subfolder share/cxx/mpi, containing mpi-related stuff
Specifically, the folder contains:

1) A small wrapper around MPI_Comm, with its rank/size stored
2) A connectivity class, used to store mesh connectivity
3) Some constexpr helpers for connectivity handling
4) A BoundaryExchange class, which takes care of exchanging contributions
   on the elements boundary (both locally and remotely).
5) Some fortran-callable routines to initialize/update/finalize the
   connectivity structure.

The connectivity structure is the only real input needed from F90, the rest
is built inside the C++ code.

The code can be a little overwhelming, due to the amount of structures introduced.
Perhaps, a good place to start to read is the banner in ConnectivityHelpers.hpp

Note: the pack/unpack phases are still serial, and needs to be
      parallelized with Kokkos