Making debugging of MPI-related error more easy.
The macro HOMMEXX_MPI_CHECK_ERROR (X) evaluates X, where X is assumed to
be an MPI function (which returns an error code, so not MPI_Wthick or
MPI_Wtime). In debug mode (when HOMMEXX_DEBUG is defined) we print a
brief message about the location of the error before aborting.