Removed dependency of Elements.hpp and Derivative.hpp on <random>
The random number generator (the engine) can be created inside the cpp file,
given the seed. So the random_init method now only needs the seed. This
removes <random> as a dependency for all the files that include Element.hpp
or Derivative.hpp.

Alternatively, if we decide we want to use generators different from the
mersenne twister, we could template random_init on the generator type.