Centralize handling of singletons, and make singleton cleanup possible.
On Cuda, in at least one unit test exe, we get the message "Kokkos::Cuda::Cuda
instance constructor : ERROR device not initialized". This is because of the
order of destruction of static variables in singleton functions.

This commit creates the class Homme::Context. A Context manages resources
previously treated as singletons. Context is meant to have two roles. First, a
Context singleton is the only singleton in the program. Second, a context need
not be a singleton, and each Context object can have different Elements,
Control, Derivative, etc., objects. (That probably isn't needed, but Context
immediately supports it.)

Finally, Context has two singleton functions: singleton(), which returns
Context&, and finalize_singleton(). The second is called in a unit test exe main
before Kokkos::finalize().