Relocating code that switches between Fortran and C++.
This code was in caar_pre_exchange_driver_mod, which now simply contains
the Fortran code of the pre exchange caar loop. Now it is inside the
prim_advance_caar_mod, and switches between Fortran and C++ for the
execution of the whole compute_and_apply_rhs subroutine.

The module caar_pre_exchange_driver_mod could be removed. However, it contains
all the subroutines of the pre-exchange loop that are used in the unit test
to check the correctness of the subkernels of CaarFunctor. Therefore, it is
hard to remove it without reducing the testing pervasiveness.