Guarding the finalization of the Context and switching to unique_ptr.
The guard is probably not needed, since calling clear() on an uninitialized
singleton is fine, but avoids the creation of a static object altogether.

The unique_ptr is not necessary either, but I can't imagine a scenario where
it makes sense to copy a Context anyways, and there was no method to expose
the shared pointers to start with.