Added testing profiles to allow selecting the dimensions of tests
The user has three choices:

(1) set HOMME_TESTING_PROFILE to one of the pre-defined profiles. These
    are: 'dev', 'short', 'nightly'.
(2) set HOMME_TESTS_NE directly (must be at least 2).
(3) do nothing.

(3) is equivalent to select 'nightly' profile (which is the default profile).
In case of (2), the profile is set to 'user-defined' and the given value of
ne is used (again, provided it is >=2). If a profile is selected, the value
of ne for that profile is used. In particular,

- 'dev': ne=2
- 'short': ne=4
- 'nightly': ne=12

The macro setTestsNE uses the name of the profile (case insensitive) to
set HOMME_TESTS_NE. This variable is in the namelists for the prtcA(B)-r0(3).nl
namelists. CMake will replace the variable with its value at configure time.