Changing typedefs of Views in Types.hpp
By explicitly saying that ViewType has a MemoryManagement template parameter,
we were not capturing the correct type of HostMirror's, which do not have a
template parameter for the MemoryManagement (it is always managed). With this
commit, we do not specify how many template arguments are passed to ViewType,
by means of a parameter pack Properties... (as it is done in Kokkos for pretty
much all View's stuff). The previous behavior is still there, when Properties...
expands to something that includes memory traits, but we now allow a list of
template arguments that does not include memory traits, as it is the case for
HostMirror views.