Changes in ConnectivityHelpers
I try to use enum classes as much as possible, with a little help from
a simple free function to convert an enum to the underlying integer, when
I need to use the label as an index in an array/view.

I also changed a little bit the constexpr arrays storing the information
on edges/corners.