Some changes in Connectivity class
The class stores a 2D view now, with 8 ConnectionInfo per element. Notice
that some of these may be MISSING, in case the element is one of the 24 that
touch one of the cube vertices.

Furthermore, I changed the ConnectionInfo and LidPos behavior. In particular,
we no longer store the 'type' of a connection (the possible types used to be
CORNER, EDGE_CONTIGUOUS_FWD, EDGE_CONTIGUOUS_BWD and EDGE_STRIDED). LidPos
does not store the position as a number [0,4), but as [0,8). In other words,
we no longer make a distinction between edge and corner, at this level. The
constexpr array CONNECTION_SIZE is used to assess how many points are on a
particular conneciton kind.