Modified the signature of the subview routines.
The old signatures were not suitable for HostMirror views, which lack the
MemoryTraits template argument (which then could not be deduced by the
compiler). We leave the memory space argument (we need it, and all views,
including HostMirror's have it) but we let other properties be optional,
with a parameter pack.