Removed the HostExecSpace and HostTeamPolicy types, and added MPIMemSpace (and views)
In the BoundaryExchange class, we now do the pack/unpack directly on device.
The MPI can then be performed on device or on host. Therefore a single
deep copy of the packed buffer is done from within the pack/unpack methods,
and no other place in the code needs a parallelization on Host (for the case
where HostSpace!=ExecSpace).

The user can set the CMake variable HOMMEXX_MPI_ON_DEVICE to enable calling
MPI routines on device pointers. Notice that if the execution space is the
host space, then this variable is irrelevant, since the behavior of the
program would be exactly the same.