Change of design for BoundaryExchange, to allow pack/unpack (and possibly MPI) on device.
The class now has 2 sets of buffers: m_send/recv_buffer and m_mpi_send/recv_buffer.
The former are always on device, and they are used for packing/unpacking.
The latter may or may not be on the device, depending on whether the user
set the CMake variable HOMMEXX_MPI_ON_DEVICE. The send/recv buffers are then
copied to/from the mpi_send/recv buffers after packing and before unpacking.
If HOMMEXX_MPI_ON_DEVICE is defined (and HostSpace is different from ExecSpace),
then this is an actual device<->host copy, otherwise it is a no-op.

WARNING: So far I tested the code on host only. I need to push to be able
         to fetch from a GPU testbed, and try it there. Bugfix commits may
         (read, will) follow.