Added user-defined copy constructor to BoundaryExchange.
The class is used as functor in pfor's. When creating a ParallelFor object,
we create a hard copy of the functor. The compiler generated implicit copy
constructor is not good enough, since it does a shallow copy. While that's
fine for Kokkos views (they have tracking), it is not ok for arrays of MPI
structures, since the clean_up method frees them. Therefore, we need to
manually copy MPI structures (or recreate them), to avoid double free's.