Reverting previous commit, plus removing functor interface from class.
There are issues with having BoundaryExchange as a functor, including its
size (when copied) and the fact that MPI structures should not be copied
(because they are opaque pointers, and we run into the risk of a double
free).

The copy-constructor solution was technically working, but opens the door
to future bugs if some additional members are added to the class and we
forget to update the copy ctor accordingly.

We now rely purely on lambda functions, making temporary copies of the class
members that are needed inside its body, to overcome the issue related to
lambda functions and the 'this' pointer.