Change in the way getters are implemented in Connectivity
I've had some issues with the #ifdef __CUDA_ARCH__ paradigm. I believe it works
only if the getter is called within a parallel region, which makes it impossible
to call the getter outside of it (and hence avoid the *this problem of lambda on
device). Therefore I switched to an explicit templated version. The used must
specify the memory space where he wants to fetch the data from. Only the
HostMemSpace and ExecMemSpace specialization are defined, of course, so other
erroneus (hopefully unintentional) calls will generate compiler errors.