Bug fix in BoundaryExchange
During the call to registration_complete, we cannot compute the offsets in
parallel, or at least not in the way we were doing. As explained in the
comment in the file, there is a reduction/scan operation over all the
connections, which would need an atomic_fetch_add. However, this may deteriorate
the concurrency of the loop, which is already not big, and given that it is
just a setup loop, I opted for the simpler host serial loop.