GPU: Call initialize_kokkos instead of Kokkos::initialize().
1. This is in place of the command-line arg --kokkos-ndevices, which is not
available when we're called from a Fortran main. (Technically, one could hack
around the command-line function calls in Fortran, but that is not great.)

A side benefit is we can insert code to get the device count automatically. So
now tests will run using all 4 GPUs.

2. Also remove Cuda-specific definitions in dimension.hpp. This is demonstrably
not needed any longer.

3. Finally, move initialize_hommexx_session to *after* the MPI_Init call, which
is the right way to do it. In particular, assignment of GPU to a rank can happen
only if MPI is alive. In fact, move it past all of the Fortran thread stuff so
Kokkos sees the final OpenMP thread count the Fortran code has decided on.