Bug fix related to kokkos and mpi initialization
Kokkos exec space must be initialized before the call to prim_init1, which
is where the C++ connectivity structure is filled. Such filling includes
the creation of some views, which in turn requires kokkos to be initialized.

So now kokkos initialization is right after MPI initialization (it CANNOT
happen before).