A few mods related to C++ mpi structures
The main reason for this is to make the initialization of C++ connectivity
more independent on each rank, eliminating some (probably confusing) code.

In particular:

- the LidPos struct is now LidGidPos, storing also the gid of the element.
- the lid in LidGidPos *may* be negative, signaling an off-process element.
- the add_connection routine now requires both lid and gid of the first and
  second element as inputs.
- the gid of the element is used to setup the send/recv tag in the mpi requests
- a negative lid is not a problem, since the lid is only needed during local
  exchanges, in which case the lid will NOT be negative.
- each process no longer needs to know the lid of the elements on other
  processes, because of what said in the previous point

The 'price' of this mod is an extra two int's to store for each connection
(the global id of the element). Shared connections do not need the lid on
the remote side, while local connection do not need the gid on either side.