Changing the way BoundaryExchange and BuffersManager are interconnected.
Instead of storing smart pointers in BM, we store raw pointers. This may
seem more dangerous at first (you cannot check whether a raw pointer is
valid, unless in the trivial case where it is null). However, we made it
so that only BE can register itself into the BM, and it will unregister
itself at destruction time. It is therefore impossible for BM to try to
access a stored dangling pointer.

This makes the setup of the two classes easier. There is no need of calling
an external function that performs the connection between BM and BE. The
only requirement is that the BM be set in the BE *before* the call to
'registration_completed()'.