Added comments, plus new featuer in BuffersManager
Added explanation of the class and its basic features/usage at top of both
BoundaryExchange.hpp and BuffersManager.hpp.

Plus, BuffersManager now stores individual needs for each customer, so
that it can avoid syncing the whole mpi buffers, and instead sync only
the part that is needed by the current customer. This could help when
the BE is added to the tracers, which have significantly more data (~10x)
transfer than the BE in the RK stages.